fix: use int instead of Decimal for Korean won amounts in API responses
Decimal(12,0) from PostgreSQL serializes as scientific notation (3.5E+5)
via Pydantic's JSON encoder. Korean won has no decimals, so int is the
correct response type — prevents confusing "3.5E+5" in API output.

+++ binjo-api/app/schemas/transaction.py
@@ -44,7 +44,8 @@ class TransactionResponse(BaseModel):
     id: str
     type: str
     category: str
-    amount: Decimal
+    # int, not Decimal — Korean won has no decimals, and Decimal serializes as "1.2E+5"
+    amount: int
     description: str | None = None
     counterparty: str | None = None
     transaction_date: date
